requirements.txt
faker
pandas
python-dotenv
groq